# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ToDo List",
        "version": "1"
    },
    "paths": {
...
        "/api/v1/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Atualizar uma tarefa existente.",
                "operationId": "9c708295c919ce91e53ecdcc78244841",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa a ser atualizada.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "abc"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "teste123"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "concluída"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Detalhes atualizados da tarefa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "abc"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "teste123"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "concluída"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarefa não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarefa não encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "ID inválido ou dados de entrada inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "ID inválido ou dados de entrada inválidos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tasks/{task}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Atualizar uma tarefa existente.",
                "description": "Atualiza uma tarefa existente do usuário autenticado na aplicação.",
                "operationId": "15c963d57e24ef5f852e46983d924425",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID da Task",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTask"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Nova descrição da tarefa"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Nova descrição da tarefa"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "concluída"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarefa não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarefa não encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro de validação dos campos."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "title": [
                                                    "O título é obrigatório."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Excluir uma tarefa existente",
                "description": "Exclui uma tarefa existente do usuário autenticado na aplicação.",
                "operationId": "bb0c584ad7034c11a8e501be038ccba5",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID da Task",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tarefa excluída com sucesso"
                    },
                    "404": {
                        "description": "Tarefa não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarefa não encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro ao excluir a tarefa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao excluir a tarefa"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UpdateTask": {
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Task 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description 1"
                    },
                    "status": {
                        "type": "string",
                        "example": "concluída"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}